slackbot/requirements.txt
slackclient
sqlalchemy ==1.4.46
psycopg2-binary
vaderSentiment
pymongo
pyjokes
python-dotenv 
pandas==1.5.3